fix: codeblock ctas
