+++ apps/masterbots.ai/components/layout/sidebar/sidebar-link.tsx
@@ -241,6 +241,7 @@ const ChatbotComponent: React.FC<ChatbotComponentProps> = React.memo(
 		const { userSlug, domain } = useParams()
 		const { setIsOpenPopup, setActiveThread } = useThread()
 
+		const isPro = routeType === 'pro'
 		const canonicalDomain = getCanonicalDomain(chatbot.name)
 		// * Default to prompt when no metadata found... Special case for BlankBot
 		const chatbotDomain = canonicalDomain || (domain as string) || 'prompt'
@@ -274,7 +275,7 @@ const ChatbotComponent: React.FC<ChatbotComponentProps> = React.memo(
 				return navigateTo({
 					urlType: 'chatbotThreadListUrl',
 					navigationParams: {
-						type: isPublic ? 'public' : 'personal',
+						type: isPro ? 'pro' : isPublic ? 'public' : 'personal',
 						category: category.name,
 						domain: chatbotDomain,
 						chatbot: chatbot.name,
@@ -303,7 +304,7 @@ const ChatbotComponent: React.FC<ChatbotComponentProps> = React.memo(
 						chatbot: chatbot?.name,
 					})
 				: urlBuilders.chatbotThreadListUrl({
-						type: isPublic ? 'public' : 'personal',
+						type: isPro ? 'pro' : isPublic ? 'public' : 'personal',
 						category: category.name,
 						domain: chatbotDomain,
 						chatbot: chatbot?.name,
