+++ apps/masterbots.ai/components/ui/codeblock.tsx
@@ -25,37 +25,57 @@ interface languageMap {
 }
 
 export const programmingLanguages: languageMap = {
-	javascript: '.js',
-	python: '.py',
-	java: '.java',
+	actionscript: '.as',
+	ada: '.ada',
+	assembly: '.asm',
+	bash: '.sh',
 	c: '.c',
-	cpp: '.cpp',
-	'c++': '.cpp',
 	'c#': '.cs',
-	ruby: '.rb',
-	php: '.php',
-	swift: '.swift',
-	'objective-c': '.m',
-	kotlin: '.kt',
-	typescript: '.ts',
-	go: '.go',
-	perl: '.pl',
-	rust: '.rs',
-	scala: '.scala',
-	haskell: '.hs',
-	lua: '.lua',
-	shell: '.sh',
-	sql: '.sql',
-	html: '.html',
-	css: '.css',
-	solidity: '.sol',
+	'c++': '.cpp',
 	cairo: '.cairo',
+	clojure: '.clj',
+	cobol: '.cob',
+	cpp: '.cpp',
+	css: '.css',
+	dart: '.dart',
+	dockerfile: '.dockerfile',
+	elixir: '.ex',
+	erlang: '.erl',
+	fortran: '.f',
+	go: '.go',
+	groovy: '.groovy',
+	haskell: '.hs',
+	html: '.html',
+	java: '.java',
+	javascript: '.js',
 	json: '.json',
-	yaml: '.yaml',
-	xml: '.xml',
+	kotlin: '.kt',
+	lua: '.lua',
 	markdown: '.md',
+	matlab: '.m',
+	'objective-c': '.m',
+	perl: '.pl',
+	php: '.php',
 	plaintext: '.txt',
+	powershell: '.ps1',
+	python: '.py',
+	r: '.r',
 	react: '.jsx',
+	ruby: '.rb',
+	rust: '.rs',
+	sass: '.sass',
+	scala: '.scala',
+	scss: '.scss',
+	shell: '.sh',
+	solidity: '.sol',
+	sql: '.sql',
+	swift: '.swift',
+	tsx: '.tsx',
+	typescript: '.ts',
+	vb: '.vb',
+	vue: '.vue',
+	xml: '.xml',
+	yaml: '.yaml',
 }
 
 const generateRandomString = (length: number, lowercase = false) => {
@@ -70,12 +90,13 @@ const generateRandomString = (length: number, lowercase = false) => {
 const CodeBlock: FC<Props> = memo(({ language, value }) => {
 	const { isCopied, copyToClipboard } = useCopyToClipboard({ timeout: 2000 })
 
-	const downloadAsFile = () => {
+	const downloadAsFile = (e: React.MouseEvent) => {
+		e.stopPropagation()
 		if (typeof window === 'undefined') {
 			return
 		}
-		const fileExtension = programmingLanguages[language] || '.file'
-		const suggestedFileName = `file-${generateRandomString(3, true)}${fileExtension}`
+		const fileExtension = programmingLanguages[language] || '.txt'
+		const suggestedFileName = `file-${generateRandomString(6, true)}${fileExtension}`
 		const fileName = window.prompt('Enter file name', suggestedFileName)
 
 		if (!fileName) {
@@ -94,13 +115,14 @@ const CodeBlock: FC<Props> = memo(({ language, value }) => {
 		URL.revokeObjectURL(url)
 	}
 
-	const onCopy = () => {
+	const onCopy = (e: React.MouseEvent) => {
+		e.stopPropagation()
 		if (isCopied) return
 		copyToClipboard(value)
 	}
 
 	return (
-		<div className="relative w-full overflow-hidden font-sans text-sm rounded-md sm:text-base">
+		<div className="relative w-full overflow-hidden font-sans text-sm rounded-md sm:text-base z-10">
 			<div
 				className={cn(
 					'flex items-center justify-between w-full bg-zinc-800 text-zinc-100',
@@ -111,7 +133,7 @@ const CodeBlock: FC<Props> = memo(({ language, value }) => {
 				<div className="flex items-center gap-0.5 sm:gap-1">
 					<Button
 						variant="ghost"
-						className="h-7 w-7 sm:h-8 sm:w-8 hover:bg-zinc-800 focus-visible:ring-1 focus-visible:ring-slate-700 focus-visible:ring-offset-0"
+						className="pointer-events-auto h-7 w-7 sm:h-8 sm:w-8 hover:bg-zinc-800 focus-visible:ring-1 focus-visible:ring-slate-700 focus-visible:ring-offset-0"
 						onClick={downloadAsFile}
 						size="icon"
 					>
@@ -121,7 +143,7 @@ const CodeBlock: FC<Props> = memo(({ language, value }) => {
 					<Button
 						variant="ghost"
 						size="icon"
-						className="h-7 w-7 sm:h-8 sm:w-8 hover:bg-zinc-800 focus-visible:ring-1 focus-visible:ring-slate-700 focus-visible:ring-offset-0"
+						className="pointer-events-auto h-7 w-7 sm:h-8 sm:w-8 hover:bg-zinc-800 focus-visible:ring-1 focus-visible:ring-slate-700 focus-visible:ring-offset-0"
 						onClick={onCopy}
 					>
 						{isCopied ? (
